+++ .meta-updater/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm-private/updater
 
+## 0.2.0
+
+### Minor Changes
+
+- b6d74c545: Allow a system's package manager to override pnpm's default settings
+
 ## 0.1.6
 
 ### Patch Changes