+++ packages/plugin-commands-listing/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/plugin-commands-listing
 
+## 4.0.19
+
+### Patch Changes
+
+- Updated dependencies [b6d74c545]
+  - @pnpm/config@13.4.0
+  - @pnpm/cli-utils@0.6.28
+
 ## 4.0.18
 
 ### Patch Changes