+++ packages/plugin-commands-setup/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/plugin-commands-setup
 
+## 1.1.12
+
+### Patch Changes
+
+- @pnpm/cli-utils@0.6.28
+
 ## 1.1.11
 
 ### Patch Changes